+++ CHANGELOG.rst
@@ -8,3 +8,11 @@ Changelog
 
    -  convert Python classes of CryptoParser to JSON (#1)
    -  add Python warepper to JSON data (#1)
+
+-  TLS
+
+   -  add capabilities of Chromium, Firefox and Opera browsers
+
+      -  `Chromium <https://en.wikipedia.org/wiki/Chromium_(web_browser)>`__
+      -  `Firefox <https://en.wikipedia.org/wiki/Firefox>`__
+      -  `Opera <https://en.wikipedia.org/wiki/Opera_(web_browser)>`__